1100805 - Fixing consumer group bind and unbind and moving tasks from tasks/consumer_group.py to consumer group cud manager
